refactor-hooks-part-deux * Rename directory of the new useRentalStation
