+++ src/components/BikeTile/BikeTile.tsx
@@ -7,7 +7,7 @@ import { TileHeader } from '../TileHeader/TileHeader'
 import { useSettings } from '../../settings/SettingsProvider'
 import { getIconColorType } from '../../utils/icon'
 import { FormFactor } from '../../../graphql-generated/mobility-v2'
-import { useRentalStations } from '../../logic/use-new-rental-stations/useRentalStations'
+import { useRentalStations } from '../../logic/use-rental-stations/useRentalStations'
 import { ErrorTile } from '../ErrorTile/ErrorTile'
 import { BikeTileRow } from './BikeTileRow'
 import classes from './BikeTile.module.scss'
