+++ src/components/BikeTile/BikeTileRow.tsx
@@ -3,7 +3,7 @@ import { Heading3 } from '@entur/typography'
 import { BicycleIcon } from '@entur/icons'
 import { getTranslation } from '../../utils/utils'
 import { WalkTrip } from '../WalkTrip/WalkTrip'
-import { RentalStation } from '../../logic/use-new-rental-stations/types'
+import { RentalStation } from '../../logic/use-rental-stations/types'
 import classes from './BikeTileRow.module.scss'
 
 interface BikeTileRowProps {
