+++ src/dashboards/Poster/components/CarTile/CarTile.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react'
 import { MobilityTile } from '../MobilityTile/MobilityTile'
-import { useRentalStations } from '../../../../logic/use-new-rental-stations/useRentalStations'
+import { useRentalStations } from '../../../../logic/use-rental-stations/useRentalStations'
 import { FormFactor } from '../../../../../graphql-generated/mobility-v2'
 import { RentalCarIcon } from '../../../../assets/icons/RentalCarIcon'
