+++ src/logic/use-rental-stations/types.ts
