+++ src/logic/use-rental-stations/useRentalStations.mobility.graphql
